tests/hardware/runner/requirements.txt
pyserial>=3.5
rich>=13.0
click>=8.0
